int8_t instead of int

+++ code/lib/Display.cpp
@@ -13,7 +13,7 @@ public:
     virtual void printData(char text1[], int8_t data, char text2[]) {}
     virtual void printData(char text1[], int8_t data1, char text2[], int8_t data2, char text3[]) {}
     virtual void printIn(char text[], int8_t x, int8_t y) {}
-    virtual void printIn(int data, int8_t x, int8_t y) {}
+    virtual void printIn(int8_t data, int8_t x, int8_t y) {}
     virtual void showClock(MyTime timeIn) {}
     virtual void blinkAt(int8_t x, int8_t y) {}
     virtual void noBlink() {}
